Extract dataplane config into a separate header file
In the following commit we will add new option to the shared
memory config, and we will need an ability to use that config in
another file. Including whole dataplane.h there will be awkward.
